+++ react-client/src/utility/utils.js
@@ -176,6 +176,7 @@ export function getVideoId (url) {
 }
 
 export function deleteFile (serverPath, handleResponse) {
+  debugger
   fetch(`${getAccountsUrl()}/deleteFile/${serverPath}`, {
     method: 'delete',
     headers: new Headers({
@@ -190,28 +191,35 @@ export function deleteFile (serverPath, handleResponse) {
   )
 }
 
-export function deleteFiles (payload, newFiles) {
+export function deleteFiles (payload, newFiles, filesToKeep) {
   debugger
-  let files = []
-  for (let i = 0; i < payload.length; i++) {
-    let component = payload[i]
-    files = files.concat(getFileIds(component))
-  }
+  let files = getFileIdsOfBroadcast
   for (let i = 0; i < files.length; i++) {
-    if (newFiles) {
-      for (let j = newFiles.length-1; j >= 0; j--) {
-        if (files[i] === newFiles[j]) {
-          newFiles.splice(j, 1)
+    let canBeDeleted = true
+    if (filesToKeep) {
+      for (let j = 0; j < filesToKeep.length; j++) {
+        if (files[i] === filesToKeep[j]) {
+          canBeDeleted = false
+          break
         }
       }
     }
-    console.log('deleting file', files[i])
-    deleteFile(files[i])
+    if (canBeDeleted) {
+      if (newFiles) {
+        for (let j = newFiles.length-1; j >= 0; j--) {
+          if (files[i] === newFiles[j]) {
+            newFiles.splice(j, 1)
+          }
+        }
+      }
+      console.log('deleting file', files[i])
+      deleteFile(files[i])
+    }
   }
   return newFiles
 }
 
-export function getFileIds (component) {
+export function getFileIdsOfComponent (component) {
   let files = []
   if (component.file && component.file.fileurl && component.file.fileurl.id) {
     files.push(component.file.fileurl.id)
@@ -223,4 +231,56 @@ export function getFileIds (component) {
     }
   }
   return files
+}
+
+export function getFileIdsOfBroadcast (payload) {
+  let files = []
+  for (let i = 0; i < payload.length; i++) {
+    let component = payload[i]
+    files = files.concat(getFileIdsOfComponent(component))
+  }
+  return files
+}
+
+
+export function getFileIdsOfMenu (menuItems) {
+  let initialFiles = []
+  for (let i = 0; i < menuItems.length; i++) {
+    let menuItem = menuItems[i]
+    if (menuItem.payload) {
+      initialFiles = initialFiles.concat(getFileIdsOfBroadcast(JSON.parse(menuItem.payload)))
+    }
+    if (menuItem.submenu) {
+      for (let j = 0; j < menuItem.submenu.length; j++) {
+        let subItem = menuItem.submenu[j]
+        if (subItem.payload) {
+          initialFiles = initialFiles.concat(getFileIdsOfBroadcast(JSON.parse(subItem.payload)))
+        }
+        if (subItem.submenu) {
+          for (let k = 0; k < subItem.submenu.length; k++) {
+            let nestedItem = subItem.submenu[k]
+            if (nestedItem.payload) {
+              initialFiles = initialFiles.concat(getFileIdsOfBroadcast(JSON.parse(nestedItem.payload)))
+            }
+          }
+        }
+      }
+    }
+  }
+  return initialFiles
+}
+
+export function deleteInitialFiles (initialFiles, currentFiles) {
+  for (let i = 0; i < initialFiles.length; i++) {
+    let foundFile = false
+    for (let j = 0; j < currentFiles.length; j++) {
+      if (initialFiles[i] === currentFiles[j]) {
+        foundFile = true
+        break
+      }
+    }
+    if (!foundFile) {
+      deleteFile(initialFiles[i])
+    }
+  }
 }
